refactor: unfiy the angentId to Long, and remove ResponseEntity in AgentController

+++ data-agent-management/src/main/java/com/alibaba/cloud/ai/dataagent/entity/AgentDatasource.java
@@ -17,14 +17,12 @@
 package com.alibaba.cloud.ai.dataagent.entity;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
-import org.springframework.format.annotation.DateTimeFormat;
-
 import java.time.LocalDateTime;
 import java.util.List;
-
+import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import lombok.AllArgsConstructor;
+import org.springframework.format.annotation.DateTimeFormat;
 
 @Data
 @NoArgsConstructor
@@ -33,7 +31,7 @@ public class AgentDatasource {
 
 	private Integer id;
 
-	private Integer agentId;
+	private Long agentId;
 
 	private Integer datasourceId;
 
@@ -53,7 +51,7 @@ public class AgentDatasource {
 	// 当前数据源选中的表
 	private List<String> selectTables;
 
-	public AgentDatasource(Integer agentId, Integer datasourceId) {
+	public AgentDatasource(Long agentId, Integer datasourceId) {
 		this.agentId = agentId;
 		this.datasourceId = datasourceId;
 		this.isActive = 1;
